change mondel and controller

+++ src/controllers/customer.controller.js
@@ -1,6 +1,7 @@
 const models = require('../models')
 const Customer = models.Customer
 const Invoice = models.Invoice
+const Users = models.Users
 const customerController = {
     create: async (req, res) => {
         const newCustomer = new Customer(req.body)
@@ -21,12 +22,15 @@ const customerController = {
         Customer.findByIdAndDelete(id , (err, customer) => {
             if(err) throw err
             if(!customer) return res.status(404).send('Customer not found')
-            customer.invoices.map(invoice => {
-                Invoice.findByIdAndDelete(invoice._id, (err, invoice) => {
-                    if(err) throw err
-                    res.send("Customer deleted => " + customer)
+            if(customer.invoices.length > 0){
+                customer.invoices.map(invoice => {
+                    Invoice.findByIdAndDelete(invoice._id, (err, invoice) => {
+                        if(err) throw err
+                        if(!invoice) return res.send('Invoice not found')
+                    })
                 })
-            })
+            }
+            res.send("Customer deleted => " + customer)
         })
     },
     getAll: async (req, res) => {
